updation

+++ ruby-on-bus/js/register.js
@@ -14,7 +14,7 @@ const showError = (field, errorText) => {
 
 const handleFormData = (e) => {
     e.preventDefault();
-    
+
     const firstnameInput = document.getElementById("firstname_input");
     const lastnameInput = document.getElementById("lastname_input");
     const emailInput = document.getElementById("email_registor_input");
@@ -23,7 +23,7 @@ const handleFormData = (e) => {
     const confirmPasswordInput = document.getElementById("confirm_password_registor_input");
     const phInput = document.getElementById("ph_input");
 
-    
+
     const firstname = firstnameInput.value
     const lastname = lastnameInput.value
     const email = emailInput.value
@@ -32,9 +32,9 @@ const handleFormData = (e) => {
     const ph = phInput.value.trim();
     const gender1 = genderInput1.value.trim();
     const gender2 = genderInput2.value.trim();
-  
+
     const emailPattern = /^[^ ]+@[^ ]+\.[a-z]{2,3}$/;
-    const  nameRegex = /^[a-zA-Z]+$/;
+    const nameRegex = /^[a-zA-Z]+$/;
     const rule = /^[a-zA-Z0-9_]+$/;
     const passwordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/;
     const phoneRegex = /^\d{10}$/;
@@ -45,28 +45,28 @@ const handleFormData = (e) => {
     if (firstname === "") {
         showError(firstnameInput, "Enter your first name");
     }
-    if(!nameRegex.test(firstname)){
+    if (!nameRegex.test(firstname)) {
         showError(firstnameInput, " Invalid name");
     }
-    if(firstname.charAt(0) === ' '){
-        showError(firstnameInput,"space is not allowed")
+    if (firstname.charAt(0) === ' ') {
+        showError(firstnameInput, "space is not allowed")
     }
-    
+
     if (lastname === "") {
         showError(lastnameInput, "Enter your last name");
     }
-    if(!nameRegex.test(lastname)){
+    if (!nameRegex.test(lastname)) {
         showError(lastnameInput, " Invalid name");
     }
-    if(lastname.charAt(0) === ' '){
-        showError(lastnameInput,"space is not allowed")
+    if (lastname.charAt(0) === ' ') {
+        showError(lastnameInput, "space is not allowed")
     }
 
     if (!emailPattern.test(email)) {
         showError(emailInput, "Enter a valid email address");
     }
-    if(email.charAt(0) === ' '){
-        showError(emailInput,"space is not allowed")
+    if (email.charAt(0) === ' ') {
+        showError(emailInput, "space is not allowed")
     }
     // if(!rule.test(email)){
     //     showError(emailInput,'spacial character not allowed');
@@ -76,8 +76,8 @@ const handleFormData = (e) => {
     if (password === "") {
         showError(passwordInput, "Enter your password");
     }
-    if(passwordRegex.test(password)){
-        showError(passwordInput,"Invalid password");
+    if (passwordRegex.test(password)) {
+        showError(passwordInput, "Invalid password");
     }
     if (confirmpassword === "") {
         showError(confirmPasswordInput, "Enter confirm password");
@@ -89,25 +89,41 @@ const handleFormData = (e) => {
     if (ph === "") {
         showError(phInput, "Enter phone number password");
     }
-    if(!phoneRegex.test(ph)){
-        showError(phInput,"Invalid phone number");
+    if (!phoneRegex.test(ph)) {
+        showError(phInput, "Invalid phone number");
     }
-    if(ph.charAt(0) === ' '){
-        showError(phInput,"space is not allowed")
+    if (ph.charAt(0) === ' ') {
+        showError(phInput, "space is not allowed")
     }
 
 
-    
-    if (gender1 === "" && gender2 === "" ) {
+    if (gender1 === "" && gender2 === "") {
         showError(genderInput1, "Select your gender");
     }
 
     const errorInputs = document.querySelectorAll(".input-box .error");
     if (errorInputs.length > 0) return;
-   let user = { "fname":firstnameInput  ,"lname":lastnameInput, "email":emailInput , "password" : passwordInput, "phone":phInput}
-   localStorage.setItem(user.email, JSON.stringify(user));
-    form.submit();
+
+    let user = { "firstname": firstname, "lastname": lastname, "email": email, "password": password, "phone": ph }
+    let Exist_user = localStorage.getItem(user.email);
+    if (Exist_user) {
+        alert("user alreday exist")
+    } else {
+        localStorage.setItem(user.email, JSON.stringify(user));
+        form.submit();
+    }
 }
 
 // Handling form submission event
 form.addEventListener("submit", handleFormData);
+
+
+// let user = localstorage.getitem(email)
+
+if (user) {
+    user = json.parse(user);
+
+}
+if (user.password === password) {
+    form.submit();
+}
